Move Save class logic to Listener\Database
Eliminate some more stuff we don't need here.
