+++ app/Model/Klass.php
@@ -5,6 +5,7 @@ use MyCLabs\Enum\Enum;
 
 class Klass extends Enum {
 	const REPO = 'Intraxia\Gistpen\Model\Repo';
+	const BLOB = 'Intraxia\Gistpen\Model\Blob';
 	const RUN = 'Intraxia\Gistpen\Model\Run';
 	const MESSAGE = 'Intraxia\Gistpen\Model\Message';
 	const STATE = 'Intraxia\Gistpen\Model\State';
